remove bootstrap (#118)
* Remove the TableWidget

This seems to have been a development step, but isn't being used
anymore.  Removing it cleans up the Bootstrap styling issues of
issue #101.

The ConnectorWidget doesn't look as good without the Bootstrap
sytles; we'll try to clean that up a bit in a later commit.

* Remove the FormWidget

This also seems to have been a development test.  It wasn't breaking
anything, but it still seems worth it to clean up.  In the process,
this tidies up imports just a little.

* Remove bootstrap dependency

* Remove files related to removed widgets

* upgrade

* fix

---------

Co-authored-by: Robert Schroll <rschroll@gmail.com>